Changed visibility type to double

+++ packages/open_meteo_api/lib/src/models/weather_current.dart
@@ -27,7 +27,7 @@ class WeatherCurrent {
   @JsonKey(name: "relative_humidity_2m")
   final int humidity;
 
-  final int visibility;
+  final double visibility;
 
   factory WeatherCurrent.fromJson(Map<String, dynamic> json) =>
       _$WeatherCurrentFromJson(json);
